+++ debussy_concert/reverse_etl/composition/clair_de_lune.py
@@ -7,9 +7,8 @@ from debussy_concert.core.motif.mixins.bigquery_job import BigQueryTimePartition
 from debussy_concert.reverse_etl.config.reverse_etl import ConfigReverseEtl
 from debussy_concert.reverse_etl.config.movement_parameters.reverse_etl import CsvFile, ReverseEtlMovementParameters
 from debussy_concert.reverse_etl.movement.reverse_etl_rdbms import ReverseEtlRdbmsMovement
-
 from debussy_concert.reverse_etl.phrase.dw_to_reverse_etl import DataWarehouseToReverseEtlPhrase
-from debussy_concert.reverse_etl.phrase.reverse_etl_temp_to_storage import DataWarehouseReverseEtlToTempToStoragePhrase
+from debussy_concert.reverse_etl.phrase.reverse_etl_to_storage import DataWarehouseReverseEtlToTempToStoragePhrase
 from debussy_concert.reverse_etl.phrase.storage_to_destination import StorageToDestinationPhrase
 from debussy_concert.reverse_etl.phrase.storage_to_rdbms_destination import StorageToRdbmsDestinationPhrase
